requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
alembic
python-dotenv
python-jose[cryptography]
passlib
python-multipart
fastapi-jwt-auth
bcrypt==4.0.1
fastapi-mail==1.4.1
slowapi
cloudinary
sphinx
pydantic-settings